+++ src/ActionMenu.hpp
@@ -13,7 +13,7 @@ private:
         bool enabled;
         const char* sprite;
         const char* id;
-        SEL_MenuHandler selector;
+        std::function<void(CCMenuItem*)> callback;
         float scale = 1.f;
     };
 
@@ -22,12 +22,6 @@ protected:
     virtual ~ActionMenu();
 
     void onScaleEnd();
-
-    void onRestart(CCObject*);
-    void onFullRestart(CCObject*);
-    void onPractice(CCObject*);
-    void onPause(CCObject*);
-    void onExit(CCObject*);
 
     void setOpacity(GLubyte opacity);
     void setScale(float scale);