+++ changelog.md
@@ -1,5 +1,6 @@
 # v1.0.2
 - Added toggle setting `Pause Level`
+- Internal optimizations
 
 # v1.0.1
 - Set minimum for number setting *`Scale`* to `0.25`